Fixed according to feedback
Schematic -> looks like the current limiting resistors for the read and blue LEDs are reversed.

Bypass caps for the MCU are drawn really weird!

Looks like you moved the XTAL and now you have two redundant wires.

PCB -> Identification is in the copper layer, not the silk layer.

C12 does not need vias, the through-hole header serves as vias.

You've got a split on the SDA line of the MCU - try to route from a via, or another pad if you can.

For the power supply, you've not used the filter caps correctly. Go from the USB connector to C8, then to regulator pin 3 (not from reg pin3 to the usb). You can put a fat track between the two pin 2s. You can add a new via size (1mm, 0.8mm) in the via drop-down, and use these vias for punching into the power planes from the caps. The regulator itself should not punch into the power plane, just it's caps should.